+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- renamed `PlatformBorder` to `PlatformRectangleBorder`
+- renamed `SquircleBorder` to `SquircleRectangleBorder`
+- added `Rectangle`, `Capsule`, and `Circle` widgets.
+- added `Rectangle.custom` for controlling individual corners 
+
 ## 0.1.0
 
 - first release